All: Re-enable CI and test runner
The followings will be left in the root of the repo:

.eslintrc.json: Configuration file for eslint, which is responsible for linting all JS scripts
Gruntfile.js: Main configuration file for grunt
infra: Folder containing all supporting scipts
package.json: Test runner definition
tasks: Folder containing grunt integration
.travis.yml: Travis CI integration
