+++ infra/lib/setup.js
