--- ocf/test-suite/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "iot-js-api-ocf-test-suite",
-  "version": "0.1.0",
-  "description": "OCF JS API Test Suite",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/01org/iot-js-api.git"
-  },
-  "keywords": [
-    "OCF",
-    "OIC",
-    "testsuite"
-  ],
-  "author": "Gabriel Schulhof <gabriel.schulhof@intel.com>",
-  "license": "Apache-2.0",
-  "bugs": {
-    "url": "https://github.com/01org/iot-js-api/issues"
-  },
-  "homepage": "https://github.com/01org/iot-js-api#readme",
-  "devDependencies": {
-    "eslint-config-jquery": "0.1.2",
-    "grunt": "^0.4.5",
-    "grunt-eslint": "18.1.0",
-    "load-grunt-config": "^0.17.2",
-    "load-grunt-tasks": "^3.2.0"
-  },
-  "dependencies": {
-    "async": "^2.0.0-rc.4",
-    "glob": "^7.0.3",
-    "lodash": "^4.12.0",
-    "qunitjs": "^2.0.0-rc1",
-    "tmp": "0.0.28",
-    "uuid": "^2.0.2"
-  }
-}
