+++ ocf/tests/oic1.1.0-0/Create - Resource (Direct)/client.js
