+++ ocf/tests/oic1.1.0-0/Delete - Resource/server.js
