+++ ocf/tests/oic1.1.0-0/Discovery - Device/client.js
