+++ ocf/tests/oic1.1.0-0/Discovery - Resource (Direct)/server.js
