+++ ocf/tests/oic1.1.0-0/Discovery - Resource (Multiple Types & Interfaces)/client.js
