+++ ocf/tests/oic1.1.0-0/Observe - Resource/client.js
