+++ ocf/tests/oic1.1.0-0/Presence/client.js
