+++ ocf/tests/oic1.1.0-0/Retrieve - Resource (Direct)/client.js
