+++ ocf/tests/oic1.1.0-0/Server - register and unregister resource.js
