+++ ocf/tests/oic1.1.0-0/Structure - Platform.js
