+++ package.json
@@ -1,13 +1,13 @@
 {
   "name": "iot-js-api",
   "description": "OCF JS API specifications and test suites",
-
-  "version": "1.0",
+  "main": "infra/index.js",
+  "version": "1.0.0",
   "versions": {
     "ocf": "oic1.1.0-0",
     "ble": "1.0",
     "board": "1.0",
-    "sensor": "1.0",
+    "sensor": "1.0"
   },
 
   "license": "Apache-2.0",
@@ -21,5 +21,21 @@
     "url": "https://github.com/01org/iot-js-api/issues"
   },
 
-  "homepage": "https://github.com/01org/iot-js-api#readme",
+  "devDependencies": {
+    "eslint-config-jquery": "0.1.2",
+    "grunt": "^0.4.5",
+    "grunt-eslint": "18.1.0",
+    "load-grunt-config": "^0.17.2",
+    "load-grunt-tasks": "^3.2.0"
+  },
+  "dependencies": {
+    "async": "^2.0.0-rc.4",
+    "glob": "^7.0.3",
+    "lodash": "^4.12.0",
+    "qunitjs": "^2.0.0-rc1",
+    "tmp": "0.0.28",
+    "uuid": "^2.0.2"
+  },
+
+  "homepage": "https://github.com/01org/iot-js-api#readme"
 }
